지렸다 리팩토링
